template_IS

A HTML template for knitted reports that matches ImmuneSpace's graphic style. It is based on html_document from the rmarkdown package with css, theme, and template parameters disabled.

template_IS(...)

Arguments

...

See html_document

Value

R Markdown output format to pass to render

Details

See the documentation for html_document or the oneline documentation for additional details on using the html_document format. Compared to html_document, it:

uses a custom css stylesheet

does not use bootstrap themes

Examples

if (FALSE) {
library(ImmuneSpaceR)
rmarkdown::render("input.Rmd", template_IS())
rmarkdown::render("input.Rmd", template_IS(toc = TRUE))
}
template_IS()
#> $knitr
#> $knitr$opts_knit
#> NULL
#> 
#> $knitr$opts_chunk
#> $knitr$opts_chunk$dev
#> [1] "png"
#> 
#> $knitr$opts_chunk$dpi
#> [1] 96
#> 
#> $knitr$opts_chunk$fig.width
#> [1] 7
#> 
#> $knitr$opts_chunk$fig.height
#> [1] 5
#> 
#> $knitr$opts_chunk$fig.retina
#> [1] 2
#> 
#> 
#> $knitr$knit_hooks
#> NULL
#> 
#> $knitr$opts_hooks
#> NULL
#> 
#> $knitr$opts_template
#> NULL
#> 
#> 
#> $pandoc
#> $pandoc$to
#> [1] "html"
#> 
#> $pandoc$from
#> [1] "markdown+autolink_bare_uris+tex_math_single_backslash"
#> 
#> $pandoc$args
#> [1] "--self-contained"                                             
#> [2] "--section-divs"                                               
#> [3] "--template"                                                   
#> [4] "/Users/runner/work/_temp/Library/rmarkdown/rmd/h/default.html"
#> [5] "--no-highlight"                                               
#> [6] "--variable"                                                   
#> [7] "highlightjs=1"                                                
#> 
#> $pandoc$keep_tex
#> [1] FALSE
#> 
#> $pandoc$latex_engine
#> [1] "pdflatex"
#> 
#> $pandoc$ext
#> NULL
#> 
#> $pandoc$convert_fun
#> NULL
#> 
#> $pandoc$lua_filters
#> [1] "/Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/pagebreak.lua"
#> [2] "/Users/runner/work/_temp/Library/rmarkdown/rmarkdown/lua/latex-div.lua"
#> 
#> 
#> $keep_md
#> [1] FALSE
#> 
#> $clean_supporting
#> [1] TRUE
#> 
#> $df_print
#> [1] "default"
#> 
#> $pre_knit
#> function (...) 
#> {
#>     op(base(...), overlay(...))
#> }
#> <bytecode: 0x7fc133787b30>
#> <environment: 0x7fc1337874d8>
#> 
#> $post_knit
#> function (...) 
#> {
#>     op(base(...), overlay(...))
#> }
#> <bytecode: 0x7fc133787b30>
#> <environment: 0x7fc133786dd8>
#> 
#> $pre_processor
#> function (...) 
#> {
#>     op(base(...), overlay(...))
#> }
#> <bytecode: 0x7fc133787b30>
#> <environment: 0x7fc1337866d8>
#> 
#> $intermediates_generator
#> function (original_input, intermediates_dir) 
#> {
#>     copy_render_intermediates(original_input, intermediates_dir, 
#>         !self_contained)
#> }
#> <bytecode: 0x7fc1151f3970>
#> <environment: 0x7fc1151f8fd0>
#> 
#> $post_processor
#> function (metadata, input_file, output_file, ...) 
#> {
#>     original_output_file <- output_file
#>     output_file <- overlay(metadata, input_file, output_file, 
#>         ...)
#>     if (!is.null(attr(output_file, "post_process_original"))) 
#>         base(metadata, input_file, original_output_file, ...)
#>     base(metadata, input_file, output_file, ...)
#> }
#> <bytecode: 0x7fc133788320>
#> <environment: 0x7fc13378d388>
#> 
#> $on_exit
#> function () 
#> {
#>     if (is.function(base)) 
#>         base()
#>     if (is.function(overlay)) 
#>         overlay()
#> }
#> <bytecode: 0x7fc13378c4e0>
#> <environment: 0x7fc13378feb0>
#> 
#> attr(,"class")
#> [1] "rmarkdown_output_format"