template_IS

A HTML template for knitted reports that matches ImmuneSpace's graphic style. It is based on html_document from the rmarkdown package with css, theme, and template parameters disabled.

template_IS(...)

Arguments

...
See html_document

Value

R Markdown output format to pass to render

Details

See the documentation for html_document or the oneline documentation for additional details on using the html_document format. Compared to html_document, it:

uses a custom css stylesheet

does not use bootstrap themes

Examples

if (FALSE) {
library(ImmuneSpaceR)
rmarkdown::render("input.Rmd", template_IS())
rmarkdown::render("input.Rmd", template_IS(toc = TRUE))
}
template_IS()

#> $knitr
#> $knitr$opts_knit
#> NULL
#> 
#> $knitr$opts_chunk
#> $knitr$opts_chunk$dev
#> [1] "png"
#> 
#> $knitr$opts_chunk$dpi
#> [1] 96
#> 
#> $knitr$opts_chunk$fig.width
#> [1] 7
#> 
#> $knitr$opts_chunk$fig.height
#> [1] 5
#> 
#> $knitr$opts_chunk$fig.retina
#> [1] 2
#> 
#> 
#> $knitr$knit_hooks
#> NULL
#> 
#> $knitr$opts_hooks
#> NULL
#> 
#> $knitr$opts_template
#> NULL
#> 
#> 
#> $pandoc
#> $pandoc$to
#> [1] "html"
#> 
#> $pandoc$from
#> [1] "markdown+autolink_bare_uris+tex_math_single_backslash+smart"
#> 
#> $pandoc$args
#>  [1] "--email-obfuscation"                                                                                                                                              
#>  [2] "none"                                                                                                                                                             
#>  [3] "--self-contained"                                                                                                                                                 
#>  [4] "--standalone"                                                                                                                                                     
#>  [5] "--section-divs"                                                                                                                                                   
#>  [6] "--template"                                                                                                                                                       
#>  [7] "/Library/Frameworks/R.framework/Versions/3.6/Resources/library/rmarkdown/rmd/h/default.html"                                                                      
#>  [8] "--no-highlight"                                                                                                                                                   
#>  [9] "--variable"                                                                                                                                                       
#> [10] "highlightjs=1"                                                                                                                                                    
#> [11] "--css"                                                                                                                                                            
#> [12] "/private/var/folders/9m/vwxpkj1100z3gpsf27m21fx00000gn/T/RtmpOlAWjV/temp_libpatha154e47eeff/ImmuneSpaceR/rmarkdown/templates/ImmuneSpace/resources/IStemplate.css"
#> 
#> $pandoc$keep_tex
#> [1] FALSE
#> 
#> $pandoc$latex_engine
#> [1] "pdflatex"
#> 
#> $pandoc$ext
#> NULL
#> 
#> 
#> $keep_md
#> [1] FALSE
#> 
#> $clean_supporting
#> [1] TRUE
#> 
#> $df_print
#> [1] "default"
#> 
#> $pre_knit
#> function (...) 
#> {
#>     op(base(...), overlay(...))
#> }
#> <bytecode: 0x7f91e3bf09c0>
#> <environment: 0x7f91e3bf12b8>
#> 
#> $post_knit
#> function (...) 
#> {
#>     op(base(...), overlay(...))
#> }
#> <bytecode: 0x7f91e3bf09c0>
#> <environment: 0x7f91e3bf19b8>
#> 
#> $pre_processor
#> function (...) 
#> {
#>     op(base(...), overlay(...))
#> }
#> <bytecode: 0x7f91e3bf09c0>
#> <environment: 0x7f91e3bf2160>
#> 
#> $intermediates_generator
#> function (original_input, intermediates_dir) 
#> {
#>     copy_render_intermediates(original_input, intermediates_dir, 
#>         !self_contained)
#> }
#> <bytecode: 0x7f91e2bfdf80>
#> <environment: 0x7f91e2bf5c00>
#> 
#> $post_processor
#> function (metadata, input_file, output_file, clean, verbose) 
#> {
#>     if (length(preserved_chunks) == 0 && !isTRUE(copy_resources) && 
#>         self_contained) 
#>         return(output_file)
#>     output_str <- read_utf8(output_file)
#>     if (length(preserved_chunks) > 0) {
#>         for (i in names(preserved_chunks)) {
#>             output_str <- gsub(paste0("<p>", i, "</p>"), i, output_str, 
#>                 fixed = TRUE, useBytes = TRUE)
#>             output_str <- gsub(paste0(" id=\"[^\"]*?", i, "[^\"]*?\" "), 
#>                 " ", output_str, useBytes = TRUE)
#>         }
#>         output_str <- restorePreserveChunks(output_str, preserved_chunks)
#>     }
#>     if (copy_resources) {
#>         output_str <- copy_html_resources(one_string(output_str), 
#>             lib_dir, output_dir)
#>     }
#>     else if (!self_contained) {
#>         image_relative <- function(img_src, src) {
#>             in_file <- utils::URLdecode(src)
#>             if (grepl("^[.][.]", in_file)) 
#>                 return(img_src)
#>             if (length(in_file) && file.exists(in_file)) {
#>                 img_src <- sub(src, utils::URLencode(normalized_relative_to(output_dir, 
#>                   in_file)), img_src, fixed = TRUE)
#>             }
#>             img_src
#>         }
#>         output_str <- process_images(output_str, image_relative)
#>     }
#>     write_utf8(output_str, output_file)
#>     output_file
#> }
#> <bytecode: 0x7f91e2bfe450>
#> <environment: 0x7f91e2bf5c00>
#> 
#> $on_exit
#> function () 
#> {
#>     if (is.function(base)) 
#>         base()
#>     if (is.function(overlay)) 
#>         overlay()
#> }
#> <bytecode: 0x7f91e12774f8>
#> <environment: 0x7f91e1277a38>
#> 
#> attr(,"class")
#> [1] "rmarkdown_output_format"